Major refactor time!
This commit refactors a lot of the functionality and call signatures.
This means that many of the classes have been renamed, and much
of the functionality has been "simplified".
