+++ superintendent/distributed/__init__.py
@@ -1,2 +1,6 @@
-from .multioutput import MultiLabeller  # noqa
-from .semisupervisor import SemiSupervisor  # noqa
+"""Interactive machine learning supervision using databases to distribute."""
+
+from .class_labeller import ClassLabeller
+from .multiclass_labeller import MultiClassLabeller
+
+__all__ = ["ClassLabeller", "MultiClassLabeller"]
