+++ superintendent/distributed/multiclass_labeller.py
@@ -1,0 +1,60 @@
+from .. import multiclass_labeller
+from .mixin import DistributedMixin
+
+
+class MultiClassLabeller(
+    DistributedMixin, multiclass_labeller.MultiClassLabeller
+):
+    """
+    A class for labelling your data.
+
+    This class is designed to label data for (semi-)supervised learning
+    algorithms. It allows you to label data. In the future, it will also allow
+    you to re-train an algorithm.
+
+    Parameters
+    ----------
+    connection_string: str
+        A SQLAlchemy-compatible database connection string. This is where the
+        data for this widget will be stored, and where it will be retrieved
+        from for labelling.
+    features : list, np.ndarray, pd.Series, pd.DataFrame, optional
+        An array or sequence of data in which each element (if 1D) or each row
+        (if 2D) represents one data point for which you'd like to generate
+        labels.
+    labels : list, np.ndarray, pd.Series, pd.DataFrame, optional
+        If you already have some labels, but would like to re-label some, then
+        you can pass these in as labels.
+    options : tuple, list
+        The options presented for labelling.
+    classifier : sklearn.base.ClassifierMixin, optional
+        An object that implements the standard sklearn fit/predict methods. If
+        provided, a button for retraining the model is shown, and the model
+        performance under k-fold crossvalidation can be read as you go along.
+    display_func : callable, optional
+        A function that will be used to display the data. This function should
+        take in two arguments, first the data to display, and second the number
+        of data points to display (set to 1 for this class).
+    eval_method : callable, optional
+        A function that accepts the classifier, features, and labels as input
+        and returns a dictionary of values that contain the key 'test_score'.
+        The default is sklearn.model_selection.cross_validate, with cv=3. Use
+        functools.partial to create a function with its parameters fixed.
+    reorder : str, callable, optional
+        One of the reordering algorithms specified in
+        :py:mod:`superintendent.prioritisation`. This describes a function that
+        receives input in the shape of n_samples, n_labels and calculates the
+        priority in terms of information value in labelling a data point.
+    shuffle_prop : float
+        The proportion of points that are shuffled when the data points are
+        re-ordered (see reorder keyword-argument). This controls the
+        "exploration vs exploitation" trade-off - the higher, the more you
+        explore the feature space randomly, the lower, the more you exploit
+        your current weak points.
+    keyboard_shortcuts : bool, optional
+        If you want to enable ipyevent-mediated keyboard capture to use the
+        keyboard rather than the mouse to submit data.
+
+    """
+
+    pass
