--- superintendent/multioutput/__init__.py
@@ -1,1 +1,0 @@
-from .multilabeller import MultiLabeller  # noqa
