+++ superintendent/queueing/__init__.py
@@ -1,0 +1,11 @@
+"""Classes for arranging queues"""
+
+from .in_memory import SimpleLabellingQueue
+from .cluster_queue import ClusterLabellingQueue
+from .base import BaseLabellingQueue
+
+__all__ = [
+    "SimpleLabellingQueue",
+    "ClusterLabellingQueue",
+    "BaseLabellingQueue",
+]
